+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`0.2.1...main`][0.2.1...main].
+For a full diff see [`0.3.0...main`][0.3.0...main].
+
+## [`0.3.0`][0.3.0]
+
+For a full diff see [`0.2.1...0.3.0`][0.2.1...0.3.0].
 
 ### Added
 
@@ -17,6 +21,7 @@ For a full diff see [`0.2.1...main`][0.2.1...main].
 
 * Moved resolution of `Count` to `FixtureFactory` ([#351]), by [@localheinz]
 * Extracted `DefaultStrategy` for resolving field values and count ([#353]), by [@localheinz]
+* Replaced `FixtureFactory::persistAfterCreate()` and `FixtureFactory::doNotPersistAfterCreate()` with `FixtureFactory::persisting()`, a mutator that returns a persisting `FixtureFactory` ([#374]), by [@localheinz]
 
 ## [`0.2.1`][0.2.1]
 
@@ -117,11 +122,13 @@ For a full diff see [`fa9c564...0.1.0`][fa9c564...0.1.0].
 [0.1.0]: https://github.com/ergebnis/factory-bot/releases/tag/0.1.0
 [0.2.0]: https://github.com/ergebnis/factory-bot/releases/tag/0.2.0
 [0.2.1]: https://github.com/ergebnis/factory-bot/releases/tag/0.2.1
+[0.3.0]: https://github.com/ergebnis/factory-bot/releases/tag/0.3.0
 
 [fa9c564...0.1.0]: https://github.com/ergebnis/factory-bot/compare/fa9c564...0.1.0
 [0.1.0...0.2.0]: https://github.com/ergebnis/factory-bot/compare/0.1.0...0.2.0
 [0.2.0...0.2.1]: https://github.com/ergebnis/factory-bot/compare/0.2.0...0.2.1
-[0.2.1...main]: https://github.com/ergebnis/factory-bot/compare/0.2.1...main
+[0.2.1...0.3.0]: https://github.com/ergebnis/factory-bot/compare/0.2.1...0.3.0
+[0.3.0...main]: https://github.com/ergebnis/factory-bot/compare/0.3.0...main
 
 [#1]: https://github.com/ergebnis/factory-bot/pull/1
 [#3]: https://github.com/ergebnis/factory-bot/pull/3
@@ -190,5 +197,6 @@ For a full diff see [`fa9c564...0.1.0`][fa9c564...0.1.0].
 [#353]: https://github.com/ergebnis/factory-bot/pull/353
 [#365]: https://github.com/ergebnis/factory-bot/pull/365
 [#369]: https://github.com/ergebnis/factory-bot/pull/369
+[#374]: https://github.com/ergebnis/factory-bot/pull/374
 
 [@localheinz]: https://github.com/localheinz